Fix the bug of confVars of long value
The ConfVars function of long value, the type should be VarType.Long

Author: hongsibao 00221460 <hongsibao@huawei.com>

Closes #153 from matadorhong/bug_ConfVar and squashes the following commits:

1125501 [hongsibao 00221460] Fix the bug of confVars of long value

+++ zeppelin-zengine/src/main/java/org/apache/zeppelin/conf/ZeppelinConfiguration.java
@@ -458,7 +458,7 @@ public class ZeppelinConfiguration extends XMLConfiguration {
       this.floatValue = -1;
       this.longValue = longValue;
       this.booleanValue = false;
-      this.type = VarType.INT;
+      this.type = VarType.LONG;
     }
 
     ConfVars(String varName, float floatValue) {
